NavigationData: Add a method to build a Map of the diff with another data
Generate a map of top-level differences between another instance.
While this could go deep in all the subtypes, we only care about the
main changes between a navigation event, so this is more than enough.

We can't just build another object of the same time, as that will not
underline the changed values (when going back to default as we want here
and that we can achieve with a map.

Changes on mutable properties are ignored on changes.