CustomSerializers: Improve serialization for Any? type to JSON primitives
When we're handling a JSON primitive type we can do the serialization
and deserialization without having to save the object type as that is
implicitly saved anyways.

Thus let's avoid it. We need to use some JsonEncoder/Decoder specific
code but this is worth given we may use it anyways for other stuff.