NavigationWebSocket: Add a service to create a Websoket Server with navigation events
The server is created using ktor implementation and coroutines to handle
the requests. For now we only support one client (as just one should be
able to control the navigation, on a first arrived, first served basis).

The events are emitted (for now) using a simple JSON encoding at every
update, as per the kotlin serialization happening in navparser.

In future we could allow also multiple listeners and one controller.